+++ package.json
@@ -40,6 +40,7 @@
     "clsx": "^2.0.0",
     "framer-motion": "^10.16.4",
     "husky": "^8.0.1",
+    "jotai": "^2.8.1",
     "million": "^2.6.4",
     "next": "14.0.2",
     "next-themes": "^0.2.1",
@@ -76,6 +77,7 @@
     "eslint-plugin-unused-imports": "^3.0.0",
     "lint-staged": "^13.2.3",
     "postcss": "^8.4.31",
+    "prettier": "^3.2.5",
     "tailwindcss": "^3.4.0",
     "typescript": "^5"
   },
