+++ src/app/(root)/state.ts
@@ -1,0 +1,34 @@
+import { atomWithReset } from 'jotai/utils'
+
+import { Video, VideoConfig } from './types'
+
+const videoConfig: VideoConfig = {
+  convertToExtension: 'mp4',
+  presetName: 'ironclad',
+  shouldDisableCompression: false,
+  shouldMuteVideo: false,
+}
+
+const video: Video = {
+  id: null,
+  isFileSelected: false,
+  pathRaw: null,
+  path: null,
+  fileName: null,
+  mimeType: null,
+  sizeInBytes: null,
+  size: null,
+  extension: null,
+  thumbnailPathRaw: null,
+  thumbnailPath: null,
+  isThumbnailGenerating: false,
+  videoDurationMilliseconds: null,
+  videDurationRaw: null,
+  isCompressing: false,
+  isCompressionSuccessful: false,
+  compressedVideo: null,
+  compressionProgress: 0,
+  config: videoConfig,
+}
+
+export const videoAtom = atomWithReset(video)
