+++ src/app/globals.css
@@ -6,12 +6,16 @@
   box-sizing: border-box;
 }
 
-* {
+*,
+*::after,
+*::before {
   -webkit-user-select: none;
   -moz-user-select: none;
   -ms-user-select: none;
   user-select: none;
   box-sizing: border-box;
+  -webkit-user-drag: none;
+  -webkit-app-region: no-drag;
 
   -webkit-overflow-scrolling: touch;
   scroll-behavior: smooth;
