+++ src/components/Icon/registry.ts
@@ -18,6 +18,8 @@ import Question from '@/assets/icons/question.svg'
 import Setting from '@/assets/icons/setting.svg'
 import Trash from '@/assets/icons/trash.svg'
 import DragAndDrop from '@/assets/icons/drag-and-drop.svg'
+import Warning from '@/assets/icons/warning.svg'
+import Error from '@/assets/icons/error.svg'
 
 type SVGAsComponent = React.FC<React.SVGProps<SVGElement>>
 
@@ -46,6 +48,8 @@ const registry = asRegistry({
   setting: Setting,
   trash: Trash,
   dragAndDrop: DragAndDrop,
+  warning: Warning,
+  error: Error,
 })
 
 export default registry
