+++ tailwind.config.ts
@@ -54,19 +54,6 @@ const config: Config = {
       },
       screens: { ...WIDTHS, ...CUSTOM_MQ },
       maxWidth: WIDTHS,
-      animation: {
-        'meteor-effect': 'meteor 5s linear infinite',
-      },
-      keyframes: {
-        meteor: {
-          '0%': { transform: 'rotate(215deg) translateX(0)', opacity: '1' },
-          '70%': { opacity: '1' },
-          '100%': {
-            transform: 'rotate(215deg) translateX(-500px)',
-            opacity: '0',
-          },
-        },
-      },
     },
   },
 }
